requirements.txt
google
beautifulsoup4
fake-useragent
transformers
torch
python-telegram-bot